Fix build and run
The current snapcraft.yaml file doesn't build because the
`license` field has an "or" statement in lowercase, when the
specification says that it must be in uppercase.

Also, the built snap returns a core dump because it tries to
use the libraries in the gnome-46-2404-sdk instead of its own
ones.

Finally, snapcraft shown several warnings due to still using
CRAFT_ARCH_TRIPLET without the _BUILD_FOR.

This PR fixes all this.